I don't think it is a good idea to store the passphrase in plaintext
Especially that the file could remain on the disk forever. Even though I
change the permission to 0600, but it does not seem to be a good idea at all.

I've tried using `https://github.com/takama/daemon`, but later on found it
requiring `sudo`. I will improve this part more specifically soon.